fix: validate signature for slack_source (#15)
I looked into writing the middleware. The middleware implementation is
trivial, but making it apply only to specific routes is less so. It
looks like fastapi currently only supports dividing middleware based on
the FastAPI app it was applied to, so you have to have a route prefix on
any set of endpoints you want to apply the middleware to without
affecting others.
e.g.
https://github.com/tiangolo/fastapi/issues/1174#issuecomment-905851468

So if we divided our core api app from the oauth endpoints, we could add
the middleware to the coreapi w/o adding it to oauth. But that's a lot
more refactoring than I want to do rn